Lab work

+++ src/sk/nixone/ds/core/time/PlannedEvent.java
@@ -8,49 +8,49 @@ public class PlannedEvent {
 	
 	@Override
 	public int compare(PlannedEvent o1, PlannedEvent o2) {
-	    return Float.compare(o1.getScheduledTime(), o2.getScheduledTime());
+	    return Double.compare(o1.getScheduledTime(), o2.getScheduledTime());
 	}
     };
     
     private Event event;
     private SimulationRun simulation;
-    private float executionTime;
-    private float scheduledTime;
+    private double executionTime;
+    private double scheduledTime;
     
-    protected PlannedEvent(SimulationRun simulation, float executionTime, float scheduledTime, Event event) {
-	this.scheduledTime = scheduledTime;
-	this.executionTime = executionTime;
-	this.event = event;
-	this.simulation = simulation;
+    protected PlannedEvent(SimulationRun simulation, double executionTime, double scheduledTime, Event event) {
+		this.scheduledTime = scheduledTime;
+		this.executionTime = executionTime;
+		this.event = event;
+		this.simulation = simulation;
     }
     
     public Event getEvent() {
-	return event;
+		return event;
     }
     
     public SimulationRun getSimulation() {
-	return simulation;
+		return simulation;
     }
     
-    public float getScheduledTime() {
-	return scheduledTime;
+    public double getScheduledTime() {
+		return scheduledTime;
     }
     
-    public float getExecutionTime() {
-	return executionTime;
+    public double getExecutionTime() {
+		return executionTime;
     }
     
     public boolean isFinished() {
-	return simulation.getCurrentSimulationTime() > executionTime;
+		return simulation.getCurrentSimulationTime() > executionTime;
     }
     
-    public float getProgress() {
-	if (isFinished()) {
-	    return 1f;
-	}
-	
-	float simulationTime = simulation.getCurrentSimulationTime();
-	
-	return (simulationTime - scheduledTime) / (executionTime - scheduledTime);
+    public double getProgress() {
+		if (isFinished()) {
+			return 1.;
+		}
+
+		double simulationTime = simulation.getCurrentSimulationTime();
+
+		return (simulationTime - scheduledTime) / (executionTime - scheduledTime);
     }
 }
