+++ src/sk/nixone/ds/core/time/RealSleepingTimeJumper.java
@@ -3,12 +3,12 @@ package sk.nixone.ds.core.time;
 public class RealSleepingTimeJumper implements TimeJumper {
 
     @Override
-    public float jump(float currentTime, float nextEventTime) {
-	try {
-	    Thread.sleep((int)((nextEventTime - currentTime)*1000));
-	} catch(InterruptedException e) {
-	    // nothing
-	}
-	return nextEventTime;
+    public double jump(double currentTime, double nextEventTime) {
+		try {
+			Thread.sleep((int)((nextEventTime - currentTime)*1000));
+		} catch(InterruptedException e) {
+			// nothing
+		}
+		return nextEventTime;
     }
 }
