+++ src/sk/nixone/ds/core/time/SimpleTimeJumper.java
@@ -3,7 +3,7 @@ package sk.nixone.ds.core.time;
 public class SimpleTimeJumper implements TimeJumper {
 
     @Override
-    public float jump(float currentTime, float nextEventTime) {
-	return nextEventTime;
+    public double jump(double currentTime, double nextEventTime) {
+		return nextEventTime;
     }
 }
