+++ src/sk/nixone/ds/core/time/TimeJumper.java
@@ -2,5 +2,5 @@ package sk.nixone.ds.core.time;
 
 public interface TimeJumper {
     
-    public float jump(float currentTime, float nextEventTime);
+    public double jump(double currentTime, double nextEventTime);
 }
